#139 refactor: 예약 목록 조회 원형에 필터 목적의 status 쿼리 파라미터 사용하도록 수정
